+++ .phan/config.php
@@ -3,6 +3,7 @@
 return [
     'directory_list' => [
         'src',
+        'vendor/guzzlehttp',
     ],
     'exclude_analysis_directory_list' => [
         'vendor/',